+++ documentation/src/test/java/net/jqwik/docs/statistics/StatisticsCoverageExamples.java
@@ -27,11 +27,14 @@ class StatisticsCoverageExamples {
 	@Property(generation = GenerationMode.RANDOMIZED)
 	void labeledStatistics(@ForAll @IntRange(min = 1, max = 10) Integer anInt) {
 		String range = anInt < 3 ? "small" : "large";
-		Statistics.label("range").collect(range);
-		Statistics.label("value").collect(anInt);
 
-		Statistics.coverageOf("range", coverage -> coverage.check("small").percentage(p -> p > 20.0));
-		Statistics.coverageOf("value", coverage -> coverage.check(0).count(c -> c > 0));
+		Statistics.label("range")
+				  .collect(range)
+				  .coverage(coverage -> coverage.check("small").percentage(p -> p > 20.0));
+
+		Statistics.label("value")
+				  .collect(anInt)
+				  .coverage(coverage -> coverage.check(0).count(c -> c > 0));
 	}
 
 	@Property